Move log module to an independent package log
